File uploaded
Script to check if SI-cards listed in check.tsv are present in the backup files of controls listed in the same file. This can be useful when checking many suspected cases of the "SI-bug" on large events. The SI-bug is a problem where a punch not always get properly registered in the card, or gets overwritten by a later station. It seems to occur about once in every 40000 or so punches with FW 656 (and earlier). The 1/40000 frequency comes from O-ringen 2019.